Optionally, keep only one video of the same playlist
If the user adds a video which is part of a playlist, all other videos
from that playlist will be deleted from the storage. Users have to
create the entry "extensions.resumelater.oneVideoPerPlaylist" as a
boolean value and set it to "true" to activate the "one video per
playlist" function.