Added videos are always on top of the list.
Even if the video was already in the list.

I think this is much better than putting the new videos to the bottom of
the list (or just keeping them on the same position as they were before
if the video was already in the list).

I think the current videos are much more important than the old ones
that the user want to resume eventually in the future. Therefore I think
it is way better if we put the videos that are important to the user to
the top of the list.

Tell me what you think about this :)