Increase minimum Firefox version to 55.0
By increasing the minimum version from 48.0 to 55.0, we are getting rid
of any warnings and notices from `web-ext lint`. These warnings and
notices were shown because we are using some APIs and permissions that
had not been available in Firefox for Android prior to version 55.0.

In theory, resumeLater should work in Firefox for Android 55.0 or later.
This has yet to be tested. Whether it works or not on Android is
probably not all that relevant, since most people are probably using the
Youtube app and not the website.

Please note that the name of the manifest key `applications` has changed
to `browser_specific_settings`. See
https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/manifest.json/browser_specific_settings
for details.